+++ clang-tools-extra/docs/ReleaseNotes.rst
@@ -97,6 +97,11 @@ Improvements to clang-tidy
 New checks
 ^^^^^^^^^^
 
+- New :doc:`avoid-passing-mlir-op-as-ref`
+  <clang-tidy/checks/llvm/avoid-passing-mlir-op-as-ref>` check.
+
+  Finds cases where ``mlir::Op`` derived classes are passed by reference.
+
 - New :doc:`llvm-type-switch-case-types
   <clang-tidy/checks/llvm/type-switch-case-types>` check.